Moved SpreadsheetHeaders to DTO folder
